Windows screen interrupt error fix
Fix for argument error to win32.event.WaitForSingleObject found on Windows 10 with Python 2.7.15.

Attempting to run samples/contact_list.py without this fix results in the following:

File "C:\Python27\lib\site-packages\asciimatics\screen.py", line 2107, in wait_for_input
    rc = win32event.WaitForSingleObject(self._stdin, timeout * 1000)
SystemError: ..\Objects\longobject.c:463: bad argument to internal function

Not sure if this is the right fix, or the best fix, but the example runs after it is applied.