Make scripts runnable in macOS and Linux (#250)
Scripts should now be runnable with `./script.py`. This uses `#!/usr/bin/env python3` so python3 just needs to be on the path. (This lets it work both in Linux as well as macOS using `brew` or other package manager.)

Added a few lines to .gitignore:
- `venv` -- for venv support
- `.eggs/` -- non-Windows `eggs/` directory.
- asciimatics/version.py -- file started out with a statement to not put it in version control, so it seemed an ommision.
- `*.log` -- At least one sample, when run, created a `.log` file. Logs are not normal things to check in.

I had these files created, but they were not checked in, and it seemed like they should never be checked in.

Closes #248